python/requirements.txt
psycopg[binary]>=3.1,<3.4
Faker>=24,<26
python-dotenv>=1.0,<2.0
